refactor: move examples functions to own directory
