+++ src/components/Parameter/OAParameterExamples.vue
@@ -1,6 +1,6 @@
 <script setup>
 import OACodeValue from '../Common/OACodeValue.vue'
-import { getPropertyExamples } from '../../lib/getPropertyExamples'
+import { getPropertyExamples } from '../../lib/examples/getPropertyExamples'
 import { useTheme } from '../../composables/useTheme'
 import OAParameterAttribute from './OAParameterAttribute.vue'
