+++ src/components/Playground/OAPlaygroundParameters.vue
@@ -8,7 +8,7 @@ import type { PlaygroundSecurityScheme, SecurityUiItem } from '../../types'
 import OAJSONEditor from '../Common/OAJSONEditor.vue'
 import OAPlaygroundParameterInput from '../Playground/OAPlaygroundParameterInput.vue'
 import OAPlaygroundSecurityInput from '../Playground/OAPlaygroundSecurityInput.vue'
-import { getPropertyExample } from '../../lib/getPropertyExample'
+import { getPropertyExample } from '../../lib/examples/getPropertyExample'
 import { buildRequest } from '../../lib/codeSamples/buildRequest'
 import { Select, SelectContent, SelectGroup, SelectItem, SelectTrigger, SelectValue } from '../ui/select'
 import type { OperationData } from '../../lib/operationData'
