+++ src/lib/examples/getPropertyExample.ts
