+++ src/lib/examples/getPropertyExamples.ts
