+++ src/lib/examples/hasExample.ts
